More tracing tool fixes than you can shake a stick at.
Now assumes all paths are in unicode, the filepaths are saved in a json string
after all, which assumes unicode strings. In theory, linux is using 'whatever
happens to be the current encoding' for its file paths, so it's technically
broken to assume UTF-8 as it is done here but it works for us(tm). And if you
are using anything else than UTF-8, you are encoding your filepaths wrong(tm).
For Windows' case, it's using the North American's Standard Unicode Encoding,
also known as, uh, CP-1252.

logman: Wrote my own CSV reader, because python's csv just doesn't do justice to
tracerpt.exe's creativity in using the same character for escaping and
quoting! This removed a lot of fragile manual quote handling code and it now
supports filepaths with comma ',' in it. And everybody knows he should write
yet another CSV file reader in 2012.

strace: Now parse strace's log C style string escape.

OSX; normalizes the unicode string to NFC*, which is the native python format,
even if OSX normalizes to NFD. Because Apple thinks using 2 unicode code points
to encode é in the file system is great.
* http://en.wikipedia.org/wiki/Unicode_equivalence

dtrace: Fixes unit tests stdout expectation to allow 10+s (!) to trace an empty
python script: thanks OSX!

Improve symlink handling, yet again. This happened to be exercised when using
$TMP on OSX has a symlink in its path, causing isolate_test_cases_smoke_test.py
to fail.

Have smoke test leaks temporary directory when using -v for easier inspection.

Added a lot of assert on the parameter type, making the whole code coherence
much simpler to assert w.r.t str vs unicode vs late-bound file paths.

R=csharp@chromium.org
BUG=


Review URL: https://chromiumcodereview.appspot.com/11360059

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/swarm_client@165711 0039d316-1c4b-4281-b951-d872f2087c98

+++ tests/isolate_test_cases_smoke_test.py
@@ -50,7 +50,12 @@ class IsolateTestCases(unittest.TestCase):
 
   def tearDown(self):
     if self.tempdir:
-      shutil.rmtree(self.tempdir)
+      if VERBOSE:
+        # If -v is used, this means the user wants to do further analisys on
+        # the data.
+        print('Leaking %s' % self.tempdir)
+      else:
+        shutil.rmtree(self.tempdir)
 
   def _copy(self, *relpath):
     relpath = os.path.join(*relpath)
@@ -131,12 +136,13 @@ class IsolateTestCases(unittest.TestCase):
     proc = subprocess.Popen(
         cmd, stdout=subprocess.PIPE, stderr=subprocess.PIPE)
     out, err = proc.communicate() or ('', '')  # pylint is confused.
+    logging.info(err)
     self.assertEqual(0, proc.returncode, (out, err))
     lines = out.splitlines()
     expected_out_re = [
-      r'\[1/3\]   \d\.\d\ds .+',
-      r'\[2/3\]   \d\.\d\ds .+',
-      r'\[3/3\]   \d\.\d\ds .+',
+      r'\[1/3\]   \d+\.\d\ds .+',
+      r'\[2/3\]   \d+\.\d\ds .+',
+      r'\[3/3\]   \d+\.\d\ds .+',
     ]
     self.assertEqual(len(expected_out_re), len(lines), (out, err))
     for index in range(len(expected_out_re)):
@@ -144,7 +150,6 @@ class IsolateTestCases(unittest.TestCase):
           re.match('^%s$' % expected_out_re[index], lines[index]),
           '%d\n%r\n%r\n%r' % (
             index, expected_out_re[index], lines[index], out))
-    logging.info(err)
     # Junk is printed on win32.
     if sys.platform != 'win32' and not VERBOSE:
       self.assertEqual('', err)
